refactor: Restructure CMake build and internal definitions
This moves a lot of files around, and changes the modulefinder
and symbolizer to be defined via compiletime definitions.
