--- src/darwin/sentry_darwin_modulefinder.h
@@ -1,8 +1,0 @@
-#ifndef SENTRY_DARWIN_MODULEFINDER_H_INCLUDED
-#define SENTRY_DARWIN_MODULEFINDER_H_INCLUDED
-
-#include "../sentry_boot.h"
-
-sentry_value_t sentry__darwin_modules_get_list(void);
-
-#endif
