+++ src/modulefinder/sentry_modulefinder_apple.c
@@ -1,6 +1,7 @@
-#include "sentry_darwin_modulefinder.h"
-#include "../sentry_sync.h"
-#include "../sentry_value.h"
+#include "sentry_boot.h"
+
+#include "sentry_sync.h"
+#include "sentry_value.h"
 
 #include <dlfcn.h>
 #include <limits.h>
@@ -125,7 +126,7 @@ done:
 }
 
 sentry_value_t
-sentry__darwin_modules_get_list(void)
+sentry__modules_get_list(void)
 {
     sentry__mutex_lock(&g_mutex);
     if (!g_initialized) {
