+++ src/path/sentry_path.c
