+++ src/path/sentry_path_windows.c
@@ -1,10 +1,10 @@
-#include "../sentry_boot.h"
+#include "sentry_boot.h"
 
-#include "../sentry_alloc.h"
-#include "../sentry_core.h"
-#include "../sentry_path.h"
-#include "../sentry_string.h"
-#include "../sentry_utils.h"
+#include "sentry_alloc.h"
+#include "sentry_core.h"
+#include "sentry_path.h"
+#include "sentry_string.h"
+#include "sentry_utils.h"
 
 #include <Pathcch.h>
 #include <Shlobj.h>
