+++ src/sentry_alloc.c
@@ -6,7 +6,7 @@
 /* on unix platforms we add support for a simplistic page allocator that can
    be enabled to make code async safe */
 #ifdef SENTRY_PLATFORM_UNIX
-#    include "unix/sentry_unix_pageallocator.h"
+#    include "sentry_unix_pageallocator.h"
 #    define WITH_PAGE_ALLOCATOR
 #endif
