+++ src/sentry_string.c
@@ -113,3 +113,16 @@ sentry__string_clonen(const char *str, size_t n)
     rv[n] = 0;
     return rv;
 }
+
+#ifdef SENTRY_PLATFORM_WINDOWS
+char *
+sentry__string_from_wstr(wchar_t *s)
+{
+    int len = WideCharToMultiByte(CP_UTF8, 0, s, -1, NULL, 0, NULL, NULL);
+    char *rv = sentry_malloc(len);
+    if (rv) {
+        WideCharToMultiByte(CP_UTF8, 0, s, -1, rv, len, NULL, NULL);
+    }
+    return rv;
+}
+#endif
