+++ src/sentry_unix_pageallocator.c
@@ -1,5 +1,5 @@
 #include "sentry_unix_pageallocator.h"
-#include "../sentry_core.h"
+#include "sentry_core.h"
 #include "sentry_unix_spinlock.h"
 
 #include <sys/mman.h>
