+++ src/sentry_windows_dbghelp.h
@@ -1,0 +1,8 @@
+#ifndef SENTRY_WINDOWS_DBGHELP_H_INCLUDED
+#define SENTRY_WINDOWS_DBGHELP_H_INCLUDED
+
+#include "sentry_boot.h"
+
+void sentry__init_dbghelp(void);
+
+#endif
