--- src/windows/sentry_windows_modulefinder.h
@@ -1,8 +1,0 @@
-#ifndef SENTRY_WINDOWS_MODULEFINDER_H_INCLUDED
-#define SENTRY_WINDOWS_MODULEFINDER_H_INCLUDED
-
-#include "../sentry_boot.h"
-
-sentry_value_t sentry__windows_modules_get_list(void);
-
-#endif
