+++ tests/sentry_testsupport.h
@@ -1,4 +1,4 @@
-#include "../src/sentry_boot.h"
+#include "sentry_boot.h"
 
 #include <inttypes.h>
 #include <setjmp.h>
