+++ tests/test_value.c
@@ -1,5 +1,5 @@
-#include "../src/sentry_value.h"
 #include "sentry_testsupport.h"
+#include "sentry_value.h"
 #include <sentry.h>
 
 SENTRY_TEST(value_null)
